sim sub-folders reorged by output
